gpu: nvgpu: therm: fix MISRA 10.3 violations
Fix MISRA Rule 10.3 violations in common/pmu/therm for assigning
objects of different size or essential type.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2027656
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/include/nvgpu/pmuif/gpmuifclk.h
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2016-2018, NVIDIA CORPORATION.  All rights reserved.
+* Copyright (c) 2016-2019, NVIDIA CORPORATION.  All rights reserved.
 *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -604,7 +604,7 @@ struct nv_pmu_clk_msg_rpc {
 };
 
 #define NV_PMU_CLK_MSG_RPC_ALLOC_OFFSET       \
-	offsetof(struct nv_pmu_clk_msg_rpc, response)
+	(u32)offsetof(struct nv_pmu_clk_msg_rpc, response)
 
 struct nv_pmu_clk_msg {
 	union {
