+++ drivers/gpu/nvgpu/include/nvgpu/pmuif/gpmuiftherm.h
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2016-2018, NVIDIA CORPORATION.  All rights reserved.
+* Copyright (c) 2016-2019, NVIDIA CORPORATION.  All rights reserved.
 *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -66,7 +66,7 @@ struct nv_pmu_therm_cmd_hw_slowdown_notification {
 };
 
 #define NV_PMU_THERM_CMD_RPC_ALLOC_OFFSET       \
-	offsetof(struct nv_pmu_therm_cmd_rpc, request)
+	(u32)offsetof(struct nv_pmu_therm_cmd_rpc, request)
 
 struct nv_pmu_therm_cmd {
 	union {
